Rename `installation.rs` -> `github.rs` now that is all it concerns
